otel: filterx file renaming
These names now follow the convention of object-string, etc...
