+++ modules/grpc/otel/filterx/object-otel-logrecord.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Attila Szakacs
+ * Copyright (c) 2023 shifter
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License version 2 as published
@@ -20,58 +20,55 @@
  *
  */
 
-#ifndef OTEL_SCOPE_HPP
-#define OTEL_SCOPE_HPP
+#ifndef OBJECT_OTEL_LOGRECORD_HPP
+#define OBJECT_OTEL_LOGRECORD_HPP
 
 #include "syslog-ng.h"
 
 #include "compat/cpp-start.h"
 #include "filterx/filterx-object.h"
-#include "otel-filterx.h"
+#include "object-otel.h"
 #include "compat/cpp-end.h"
 
-#include "opentelemetry/proto/common/v1/common.pb.h"
+#include "opentelemetry/proto/logs/v1/logs.pb.h"
 
-typedef struct FilterXOtelScope_ FilterXOtelScope;
+typedef struct FilterXOtelLogRecord_ FilterXOtelLogRecord;
 
-FilterXObject *_filterx_otel_scope_clone(FilterXObject *s);
+FilterXObject *_filterx_otel_logrecord_clone(FilterXObject *s);
 
 namespace syslogng {
 namespace grpc {
 namespace otel {
-namespace filterx {
 
-class Scope
+using opentelemetry::proto::logs::v1::LogRecord;
+
+class OtelLogRecordCpp
 {
 public:
-  Scope(FilterXOtelScope *s);
-  Scope(FilterXOtelScope *s, FilterXObject *protobuf_object);
-  Scope(Scope &o) = delete;
-  Scope(Scope &&o) = delete;
-
-  std::string marshal();
-  bool set_field(const gchar *attribute, FilterXObject *value);
-  FilterXObject *get_field(const gchar *attribute);
-  const opentelemetry::proto::common::v1::InstrumentationScope &get_value() const;
-
+  OtelLogRecordCpp(FilterXOtelLogRecord *folr);
+  OtelLogRecordCpp(FilterXOtelLogRecord *folr, FilterXObject *protobuf_object);
+  OtelLogRecordCpp(OtelLogRecordCpp &o) = delete;
+  OtelLogRecordCpp(OtelLogRecordCpp &&o) = delete;
+  FilterXObject *FilterX();
+  bool SetField(const gchar *attribute, FilterXObject *value);
+  std::string Marshal(void);
+  FilterXObject *GetField(const gchar *attribute);
+  const LogRecord &GetValue() const;
 private:
-  Scope(const Scope &o, FilterXOtelScope *s);
-  friend FilterXObject *::_filterx_otel_scope_clone(FilterXObject *s);
-
-private:
-  FilterXOtelScope *super;
-  opentelemetry::proto::common::v1::InstrumentationScope scope;
+  FilterXOtelLogRecord *super;
+  LogRecord logRecord;
+  OtelLogRecordCpp(const OtelLogRecordCpp &o, FilterXOtelLogRecord *folr);
+  friend FilterXObject *::_filterx_otel_logrecord_clone(FilterXObject *s);
 };
 
 }
 }
 }
-}
 
-struct FilterXOtelScope_
+struct FilterXOtelLogRecord_
 {
   FilterXObject super;
-  syslogng::grpc::otel::filterx::Scope *cpp;
+  syslogng::grpc::otel::OtelLogRecordCpp *cpp;
 };
 
 #endif
