+++ modules/grpc/otel/filterx/object-otel-resource.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 shifter
+ * Copyright (c) 2024 Attila Szakacs
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License version 2 as published
@@ -20,55 +20,58 @@
  *
  */
 
-#ifndef OTEL_LOG_RECORD_HPP
-#define OTEL_LOG_RECORD_HPP
+#ifndef OBJECT_OTEL_RESOURCE_HPP
+#define OBJECT_OTEL_RESOURCE_HPP
 
 #include "syslog-ng.h"
 
 #include "compat/cpp-start.h"
 #include "filterx/filterx-object.h"
-#include "otel-filterx.h"
+#include "object-otel.h"
 #include "compat/cpp-end.h"
 
-#include "opentelemetry/proto/logs/v1/logs.pb.h"
+#include "opentelemetry/proto/resource/v1/resource.pb.h"
 
-typedef struct FilterXOtelLogRecord_ FilterXOtelLogRecord;
+typedef struct FilterXOtelResource_ FilterXOtelResource;
 
-FilterXObject *_filterx_otel_logrecord_clone(FilterXObject *s);
+FilterXObject *_filterx_otel_resource_clone(FilterXObject *s);
 
 namespace syslogng {
 namespace grpc {
 namespace otel {
+namespace filterx {
 
-using opentelemetry::proto::logs::v1::LogRecord;
-
-class OtelLogRecordCpp
+class Resource
 {
 public:
-  OtelLogRecordCpp(FilterXOtelLogRecord *folr);
-  OtelLogRecordCpp(FilterXOtelLogRecord *folr, FilterXObject *protobuf_object);
-  OtelLogRecordCpp(OtelLogRecordCpp &o) = delete;
-  OtelLogRecordCpp(OtelLogRecordCpp &&o) = delete;
-  FilterXObject *FilterX();
-  bool SetField(const gchar *attribute, FilterXObject *value);
-  std::string Marshal(void);
-  FilterXObject *GetField(const gchar *attribute);
-  const LogRecord &GetValue() const;
+  Resource(FilterXOtelResource *s);
+  Resource(FilterXOtelResource *s, FilterXObject *protobuf_object);
+  Resource(Resource &o) = delete;
+  Resource(Resource &&o) = delete;
+
+  std::string marshal();
+  bool set_field(const gchar *attribute, FilterXObject *value);
+  FilterXObject *get_field(const gchar *attribute);
+  const opentelemetry::proto::resource::v1::Resource &get_value() const;
+
 private:
-  FilterXOtelLogRecord *super;
-  LogRecord logRecord;
-  OtelLogRecordCpp(const OtelLogRecordCpp &o, FilterXOtelLogRecord *folr);
-  friend FilterXObject *::_filterx_otel_logrecord_clone(FilterXObject *s);
+  Resource(const Resource &o, FilterXOtelResource *s);
+  friend FilterXObject *::_filterx_otel_resource_clone(FilterXObject *s);
+
+private:
+  FilterXOtelResource *super;
+  opentelemetry::proto::resource::v1::Resource resource;
 };
 
 }
 }
 }
+}
 
-struct FilterXOtelLogRecord_
+struct FilterXOtelResource_
 {
   FilterXObject super;
-  syslogng::grpc::otel::OtelLogRecordCpp *cpp;
+  syslogng::grpc::otel::filterx::Resource *cpp;
 };
 
 #endif
